fix(api): improve OpenAPI spec compliance with 'null' support for optional arguments and correctly documented response codes
fix(api): fix various incorrectly unhandled database errors, including no max integer size check or joining tables incorrectly

+++ api/routers/v1/courses.py
@@ -7,6 +7,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 
 from api.models import Course
 from api.util.db import aget_session
+from api.util.pydantic_type import Int64
 
 tracer = trace.get_tracer(__name__)
 
@@ -16,7 +17,7 @@ router = APIRouter(prefix="/course", tags=["Courses"])
 @router.get("/get/{unit_id}", response_model=list[Course])
 async def get_courses(
     session: Annotated[AsyncSession, Depends(aget_session)],
-    unit_id: int,
+    unit_id: Int64,
 ) -> Sequence[Course]:
     with tracer.start_as_current_span("get_courses") as span:
         span.set_attribute("unit_id", unit_id)
